unify clipping in base_layer
