+++ napari/_vispy/vispy_shapes_layer.py
@@ -26,9 +26,6 @@ class VispyShapesLayer(VispyBaseLayer):
             self._on_text_change, self._on_blending_change
         )
         self.layer.events.highlight.connect(self._on_highlight_change)
-        self.layer.experimental_clipping_planes.events.connect(
-            self._on_experimental_clipping_planes_change
-        )
 
         self.reset()
         self._on_data_change()
@@ -161,16 +158,10 @@ class VispyShapesLayer(VispyBaseLayer):
         text_node.set_gl_state(str(self.layer.text.blending))
         self.node.update()
 
-    def _on_experimental_clipping_planes_change(self, event=None):
-        self.node.clipping_planes = (
-            self.layer.experimental_clipping_planes.as_array()
-        )
-
     def reset(self):
         self._reset_base()
         self._on_highlight_change()
         self._on_blending_change()
-        self._on_experimental_clipping_planes_change()
 
     def close(self):
         """Vispy visual is closing."""
