+++ napari/layers/points/points.py
@@ -25,7 +25,6 @@ from ..utils.layer_utils import (
     get_current_properties,
     prepare_properties,
 )
-from ..utils.plane import ClippingPlane, ClippingPlaneList
 from ..utils.text_manager import TextManager
 from ._points_constants import SYMBOL_ALIAS, Mode, Symbol
 from ._points_mouse_bindings import add, highlight, select
@@ -289,6 +288,7 @@ class Points(Layer):
             opacity=opacity,
             blending=blending,
             visible=visible,
+            experimental_clipping_planes=experimental_clipping_planes,
         )
 
         self.events.add(
@@ -367,7 +367,6 @@ class Points(Layer):
         self._is_selecting = False
         self._clipboard = {}
         self._round_index = False
-        self._experimental_clipping_planes = ClippingPlaneList()
 
         self._edge = ColorManager._from_layer_kwargs(
             n_colors=len(data),
@@ -395,8 +394,6 @@ class Points(Layer):
         self.current_properties = get_current_properties(
             self._properties, self._property_choices, len(self.data)
         )
-
-        self.experimental_clipping_planes = experimental_clipping_planes
 
         # Trigger generation of view slice and thumbnail
         self._update_dims()
@@ -675,21 +672,6 @@ class Points(Layer):
                 self.size[i, :] = (self.size[i, :] > 0) * size
             self.refresh()
             self.events.size()
-
-    @property
-    def experimental_clipping_planes(self):
-        return self._experimental_clipping_planes
-
-    @experimental_clipping_planes.setter
-    def experimental_clipping_planes(
-        self, value: Union[List[Union[ClippingPlane, dict]], ClippingPlaneList]
-    ):
-        self._experimental_clipping_planes.clear()
-        if value is not None:
-            for new_plane in value:
-                plane = ClippingPlane()
-                plane.update(new_plane)
-                self._experimental_clipping_planes.append(plane)
 
     @property
     def edge_width(self) -> Union[None, int, float]:
@@ -992,9 +974,6 @@ class Points(Layer):
                 'size': self.size,
                 'ndim': self.ndim,
                 'data': self.data,
-                'experimental_clipping_planes': [
-                    plane.dict() for plane in self.experimental_clipping_planes
-                ],
             }
         )
         return state
